Move LogInfo, LogError to utils
